requirements.txt
alembic==1.14.1
blinker==1.9.0
cachelib==0.9.0
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
filetype==1.2.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-paginate==2024.4.12
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
Pygments==2.19.1
PyMySQL==1.1.1
python-dotenv==1.0.1
rich==13.9.4
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
